Try #1 at debugging
For the models, we will be using the X_features. As you can see, I also created a X_features_2d. This is for plotting a subject's connectome as heatmap (64x64) or for other purposes,  but the main goal for this is to not have to reshape or reload the data functional data again if we ever need it in that form. 

See if runs ok.